+++ NEWS.md
@@ -1,3 +1,10 @@
+## mbmixture 0.2-1 (2018-09-06)
+
+- Added function `bootstrapNull()` which is like `bootstrapSE()` but
+  simulates under the NULL model (of no contamination) and gets the
+  LRT statistics vs p=0.
+
+
 ## mbmixture 0.1-1 (2018-08-11)
 
 - A new package.